sensible data from sensor, proper accelerometer calibration. acc + mag roll, pitch and yaw calculation

+++ Commons/DataFrame.h
@@ -112,11 +112,11 @@ typedef struct
 
 typedef struct
 {
-	uint8_t compass;
+	float compass;
 
-	uint8_t pitch;
-	uint8_t yaw;
-	uint8_t roll;
+	double pitch;
+	double yaw;
+	double roll;
 
 	float accelx;
 	float accely;
